Small debugs, adding some tests
Couldn't quite get the RGB camera to work, but still great progress this week. It is likely something to do with the video output not successfully closing, or the program looping too fast so the frame can't be processed (since it is so big). If that is the case, I will probably make another thread and pass the frame along to the video writer. Regardless though, the thermal camera data and video output is functional, which is the most important part