Fixing dependencies & outlining bounding box counting
Fixing requirements.txt to refer to the correct versions, fixing errors when installing. Also starting to build the pipeline for processing bounding boxes generated by the YOLO model. Once the image to thermal data conversion is completed and everything is tested and optimized, we can begin implementing feature extraction to be processed by the model.