Adding parameter grid tests and many visualizations
Adding a parameter grid search for the YOLO model, searching many parameters (around 300 cases) to find the truly optimal fit. Additionally, since this is going to take a very long time to execute, adding a ton of different visualizations/analyses to be able to paint a full picture of the results. We should be able to determine everything that we need to know from this and gauge a clear direction of what the next steps would be.