Retraining RF Model, Optimizing YOLO params
Retraining the random forest model on the new labeled data with more parameters, saving all under their respective `.pkl` files. Additionally, redoing the YOLO parameter grid search to attempt to optimize. Received around 99.6%, which is sweet but should be taken with a grain of salt considering the limited training data.